+++ functions/scripts/sendPushCanaryForUser.js
@@ -178,7 +178,7 @@ async function main() {
     },
     webpush: {
       fcmOptions: {
-        link: "https://www.setlistpickem.com/dashboard/notifications",
+        link: "https://www.setlistpickem.com/dashboard/profile/notifications",
       },
     },
   });